refactor: [tooltip] switch to script-setup syntax
