+++ packages/components/tooltip/index.ts
@@ -3,5 +3,5 @@ import Tooltip from './src/tooltip.vue'
 
 export const ElTooltip = withInstall(Tooltip)
 export * from './src/tooltip'
-export * from './src/tokens'
+export * from '@element-plus/tokens/tooltip'
 export default ElTooltip
