+++ packages/components/tooltip/src/tooltip.vue
@@ -47,181 +47,45 @@
   </el-popper>
 </template>
 
-<script lang="ts">
-import {
-  computed,
-  defineComponent,
-  onDeactivated,
-  provide,
-  readonly,
-  ref,
-  toRef,
-  unref,
-  watch,
-} from 'vue'
-import {
-  ElPopper,
-  ElPopperArrow,
-  usePopperArrowProps,
-  usePopperProps,
-} from '@element-plus/components/popper'
-
-import { debugWarn, isBoolean, isUndefined } from '@element-plus/utils'
-import {
-  createModelToggleComposable,
-  useDelayedToggle,
-  useId,
-  usePopperContainer,
-} from '@element-plus/hooks'
+<script lang="ts" setup>
+import { ElPopper, ElPopperArrow } from '@element-plus/components/popper'
 import ElTooltipContent from './content.vue'
 import ElTooltipTrigger from './trigger.vue'
-import {
-  useTooltipContentProps,
-  useTooltipProps,
-  useTooltipTriggerProps,
-} from './tooltip'
-import { TOOLTIP_INJECTION_KEY } from './tokens'
+import { useTooltipEmits, useTooltipProps } from './tooltip'
+import { useTooltip } from './use-tooltip'
 
-const { useModelToggleProps, useModelToggle, useModelToggleEmits } =
-  createModelToggleComposable('visible' as const)
-
-export default defineComponent({
+defineOptions({
   name: 'ElTooltip',
-  components: {
-    ElPopper,
-    ElPopperArrow,
-    ElTooltipContent,
-    ElTooltipTrigger,
-  },
-  props: {
-    ...usePopperProps,
-    ...useModelToggleProps,
-    ...useTooltipContentProps,
-    ...useTooltipTriggerProps,
-    ...usePopperArrowProps,
-    ...useTooltipProps,
-  },
-  emits: [
-    ...useModelToggleEmits,
-    'before-show',
-    'before-hide',
-    'show',
-    'hide',
-    'open',
-    'close',
-  ],
-  setup(props, { emit }) {
-    usePopperContainer()
-    const compatShowAfter = computed(() => {
-      if (!isUndefined(props.openDelay)) {
-        debugWarn(
-          'ElTooltip',
-          'open-delay is about to be deprecated in the next major version, please use `show-after` instead'
-        )
-      }
-      return props.openDelay || (props.showAfter as number)
-    })
-    const compatShowArrow = computed(() => {
-      if (!isUndefined(props.visibleArrow)) {
-        debugWarn(
-          'ElTooltip',
-          '`visible-arrow` is about to be deprecated in the next major version, please use `show-arrow` instead'
-        )
-      }
-      return isBoolean(props.visibleArrow)
-        ? props.visibleArrow
-        : props.showArrow
-    })
+})
 
-    const id = useId()
-    const popperRef = ref<InstanceType<typeof ElPopper> | null>(null)
-    const contentRef = ref<InstanceType<typeof ElTooltipContent> | null>(null)
+const props = defineProps(useTooltipProps)
 
-    const updatePopper = () => {
-      const popperComponent = unref(popperRef)
-      if (popperComponent) {
-        popperComponent.popperInstanceRef?.update()
-      }
-    }
-    const open = ref(false)
-    const toggleReason = ref<Event | undefined>(undefined)
+const emit = defineEmits(useTooltipEmits)
 
-    const { show, hide } = useModelToggle({
-      indicator: open,
-      toggleReason,
-    })
+const {
+  compatShowAfter,
+  compatShowArrow,
+  popperRef,
+  contentRef,
+  open,
+  hide,
+  isFocusInsideContent,
+  updatePopper,
+  onOpen,
+  onClose,
+} = useTooltip(props, emit)
 
-    const { onOpen, onClose } = useDelayedToggle({
-      showAfter: compatShowAfter,
-      hideAfter: toRef(props, 'hideAfter'),
-      open: show,
-      close: hide,
-    })
-
-    const controlled = computed(() => isBoolean(props.visible))
-
-    provide(TOOLTIP_INJECTION_KEY, {
-      controlled,
-      id,
-      open: readonly(open),
-      trigger: toRef(props, 'trigger'),
-      onOpen: (event?: Event) => {
-        onOpen(event)
-      },
-      onClose: (event?: Event) => {
-        onClose(event)
-      },
-      onToggle: (event?: Event) => {
-        if (unref(open)) {
-          onClose(event)
-        } else {
-          onOpen(event)
-        }
-      },
-      onShow: () => {
-        emit('show', toggleReason.value)
-      },
-      onHide: () => {
-        emit('hide', toggleReason.value)
-      },
-      onBeforeShow: () => {
-        emit('before-show', toggleReason.value)
-      },
-      onBeforeHide: () => {
-        emit('before-hide', toggleReason.value)
-      },
-      updatePopper,
-    })
-
-    watch(
-      () => props.disabled,
-      (disabled) => {
-        if (disabled && open.value) {
-          open.value = false
-        }
-      }
-    )
-
-    const isFocusInsideContent = () => {
-      const popperContent: HTMLElement | undefined =
-        contentRef.value?.contentRef?.popperContentRef
-      return popperContent && popperContent.contains(document.activeElement)
-    }
-
-    onDeactivated(() => open.value && hide())
-
-    return {
-      compatShowAfter,
-      compatShowArrow,
-      popperRef,
-      contentRef,
-      open,
-      hide,
-      isFocusInsideContent,
-      updatePopper,
-      onOpen,
-      onClose,
-    }
-  },
+defineExpose({
+  disabled: props.disabled,
+  compatShowAfter,
+  compatShowArrow,
+  popperRef,
+  contentRef,
+  open,
+  hide,
+  isFocusInsideContent,
+  updatePopper,
+  onOpen,
+  onClose,
 })
 </script>
