+++ packages/tokens/index.ts
@@ -16,3 +16,4 @@ export * from './experimental-features'
 export * from './popper'
 export * from './tooltip-v2'
 export * from './date-picker'
+export * from './tooltip'
